fake-news-checker/backend/requirements.txt
# === Core Framework ===
fastapi==0.115.0 
uvicorn[standard]==0.30.0 
pydantic==2.9.0 

# === Web Scraping & Search ===
beautifulsoup4==4.12.3 
requests==2.32.3 
curl_cffi 

# === NLP & ML - Core ===
torch>=2.0.0 
transformers>=4.30.0 
sentence-transformers==3.3.0 
numpy==1.26.4 

# === Vietnamese NLP - PhoBERT Support ===
keybert>=0.7.0 
underthesea>=6.7.0 

# === Utilities ===
python-dotenv==1.0.1